requirements.txt
tensorflow          # Treinar a IA
keras               # Incluído no TensorFlow
numpy               # Para operações numéricas
matplotlib          # Para plotar gráficos de treino
Pillow              # Para manipulação de imagens
Flask
Flask_CORS
pymongo
opencv-python
rembg
ultralytics